% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfi_ggm.R
\name{plot.dfi_ggm}
\alias{plot.dfi_ggm}
\title{The invariance partial pruning (IVPP) algorithm for panel GVAR models}
\usage{
\method{plot}{dfi_ggm}()
}
\arguments{
\item{net}{The empirical network to estimate dynamic fit index cutoffs for}

\item{power}{The power that the cutoff value aims for}

\item{iter}{The number of iterations used in your simulation.}

\item{n}{Your sample size}

\item{propPos}{Decide the proportion of positive in the extra edges added to the empirical network to create the mis-specified network}

\item{ordinal}{Logical; should ordinal data be generated?}

\item{nLevels}{Number of levels used in ordinal data.}

\item{skewFactor}{How skewed should ordinal data be? 1 indicates uniform data and higher values increase skewedness.}

\item{type}{Should thresholds for ordinal data be sampled at random or determined uniformly?}

\item{missing}{Proportion of data that should be simulated to be missing.}

\item{ncores}{How many cores you want to use in the simulation. Recommend to leave one core free so that other tasks in the system are not impacted.}

\item{n_misspec}{Number of mis-specified model you want in the simulation. Default to 5, meaning
there are five mis-specified models with 1, 2, ..., 5 extra edges respectively.
Avoid setting too large numbers. Otherwise the simulation might fail or take too long.}
}
\value{
An object of class dfi_ggm. Can use summary to view a summary of results
}
\description{
Plot the fit distribution
}
