% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfi_ggm.R
\name{dfi_ggm}
\alias{dfi_ggm}
\title{The invariance partial pruning (IVPP) algorithm for panel GVAR models}
\usage{
dfi_ggm(
  net,
  power = 0.8,
  iter = 200,
  n = 500,
  propPos = 0.8,
  ordinal = FALSE,
  nLevels = 4,
  skewFactor = 1,
  type = c("uniform", "random"),
  missing = 0
)
}
\arguments{
\item{net}{The empirical network to estimate dynamic fit index cutoffs for}

\item{power}{The power that the cutoff value aims for}

\item{iter}{The number of iterations used in your simulation.}

\item{n}{Your sample size}

\item{propPos}{Decide the proportion of positive in the extra edges added to the empirical network to create the mis-specified network}

\item{ordinal}{Logical; should ordinal data be generated?}

\item{nLevels}{Number of levels used in ordinal data.}

\item{skewFactor}{How skewed should ordinal data be? 1 indicates uniform data and higher values increase skewedness.}

\item{type}{Should thresholds for ordinal data be sampled at random or determined uniformly?}

\item{missing}{Proportion of data that should be simulated to be missing.}
}
\value{
An object of class dfi_ggm. Can use summary to view a summary of results
}
\description{
This function implements the IVPP algorithm to compare networks in the multi-group panelGVAR models.
The IVPP algorithm is a two-step procedure that first conducts an global invariance test of network difference and then performs partial pruning for the specific edge-level differences.
Currently supports the comparison of temporal and contemporaneous networks.
}
