% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netDFI-package.R
\docType{package}
\name{netDFI-package}
\alias{netDFI}
\alias{netDFI-package}
\title{netDFI: Dynamic Fit Index for Network Models}
\description{
Implement methods to determine the dynamic fit index cutoffs for network models. The package allows users to evaluate model fit based on their own model statement, model type, and sample size.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/xinkaidupsy/netDFI}
  \item Report bugs at \url{https://github.com/xinkaidupsy/netDFI/issues}
}

}
\author{
\strong{Maintainer}: Xinkai Du \email{xinkai.du.xd@gmail.com} (\href{https://orcid.org/0000-0002-4158-7878}{ORCID})

}
\keyword{internal}
