% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_plot.r
\name{plot.dfi_ggm}
\alias{plot.dfi_ggm}
\title{Plot functions}
\usage{
\method{plot}{dfi_ggm}(x, ...)
}
\arguments{
\item{x}{an object of class dfi_ggm}

\item{...}{further arguments passed from plot(); currently ignored.}
}
\value{
Plots
}
\description{
Plot the fit distribution
}
